requirements.txt
anthropic>=0.40.0
tavily-python>=0.5.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn>=0.32.0
pywebview>=5.0.0
watchdog>=6.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
Pillow>=11.0.0
pydantic>=2.10.0
qrcode>=8.0
python-multipart>=0.0.9
